HAIS-software/test/requirements.txt
##### Deep learning packages
#pytorch 
torch
torchvision
tensorflow
torchviz
albumentations
ensemble-boxes
tensorboard

############ Firebase packages
pyrebase
requests>=2.26
scp
pynput

######## Web API packages
fastapi 
uvicorn
python-multipart
Pyrebase
Jinja2
# Flask

########  MySQL
mysql-connector-python
mysql-connector-python-rf

##### preprocessing packages
Pillow
# opencv-python
# opencv-python-headless 
# tifffile
sklearn
scikit-image>=0.19.1
imutils
pylops

##### Loading data packages
pynrrd
nibabel

##### GUI packages
PyQt5
napari[pyqt5]
magicgui
multiprocess
fury

############ Visualization packages
osmnx==1.1.2
folium
scikit-learn 
geopandas


############ Annotation packages
# nuscens
nuscenes-devkit==1.1.9
#UltimateLabeling
opencv-contrib-python==4.1.2.30
paramiko
# folium
# labelme Annotation packages
video-cli

##### General packages
PyYAML
matplotlib
numpy==1.21
pandas
scipy
PyYAML
glob2
progressbar
tqdm
seaborn
multiprocess

# testing code package
pytest


